requirements.txt
bottle==0.12.25
branca==0.6.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
customtkinter==5.2.0
darkdetect==0.8.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
Pillow==10.0.0
proxy-tools==0.1.0
pyperclip==1.8.2
python-decouple==3.8
pywebview==4.2.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tkintermapview==1.29
tkintertable==1.3.3
typing_extensions==4.7.1
urllib3==2.0.4
